Moved some classes around.
