+++ common/src/main/java/com/collective/projectcore/blocks/CoreBlock.java
@@ -1,0 +1,10 @@
+package com.collective.projectcore.blocks;
+
+import net.minecraft.block.Block;
+
+public class CoreBlock extends Block {
+
+    public CoreBlock(Settings settings) {
+        super(settings);
+    }
+}
