+++ common/src/main/java/com/collective/projectcore/blocks/CoreBlockWithEntity.java
@@ -1,0 +1,11 @@
+package com.collective.projectcore.blocks;
+
+import net.minecraft.block.BlockWithEntity;
+
+public abstract class CoreBlockWithEntity extends BlockWithEntity {
+
+    protected CoreBlockWithEntity(Settings settings) {
+        super(settings);
+    }
+
+}
