--- common/src/main/java/com/collective/projectcore/blocks/base/CoreBaseBlock.java
@@ -1,10 +1,0 @@
-package com.collective.projectcore.blocks.base;
-
-import net.minecraft.block.Block;
-
-public class CoreBaseBlock extends Block {
-
-    public CoreBaseBlock(Settings settings) {
-        super(settings);
-    }
-}
