--- common/src/main/java/com/collective/projectcore/blocks/base/CoreBaseBlockWithEntity.java
@@ -1,11 +1,0 @@
-package com.collective.projectcore.blocks.base;
-
-import net.minecraft.block.BlockWithEntity;
-
-public abstract class CoreBaseBlockWithEntity extends BlockWithEntity {
-
-    protected CoreBaseBlockWithEntity(Settings settings) {
-        super(settings);
-    }
-
-}
